local-agents/templates/django/requirements.txt
django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.5
django-celery-results==2.5.1

# Testing
pytest==8.2.0
pytest-django==4.8.0

# Dev tooling
ruff==0.4.7
